minor refactor, created new models for ReadGoal and UserPreferences
